+++ src/features/shows/features/my-shows/components/MyShowsContent.tsx
@@ -1,12 +1,20 @@
 import { useContext } from 'react';
+import { Box } from '@mui/material';
+import ListIcon from '@mui/icons-material/List';
 import { MyShowsContext } from '../contexts/MyShowsContext';
+import ShowsCollection from '../../../components/ShowsCollection';
+import Section from '../../../../../components/Section';
 
 const MyShowsContent = () => {
-  const { shows } = useContext(MyShowsContext);
+  const { shows, loading } = useContext(MyShowsContext);
 
-  console.log('shows', shows);
-
-  return <div>hello</div>;
+  return (
+    <Box>
+      <Section title="My Shows" icon={<ListIcon />}>
+        <ShowsCollection shows={shows} loading={loading} scroll={false} />
+      </Section>
+    </Box>
+  );
 };
 
 export default MyShowsContent;
