+++ src/features/shows/features/similar/components/SimilarShowsSection.tsx
@@ -1,14 +1,14 @@
 import React, { useContext } from 'react';
 import Section from '../../../../../components/Section';
 import { SimilarShowsContext } from '../contexts/SimilarShowsContext';
-import SimilarShowsCollection from './SimilarShowsCollection';
+import ShowsCollection from '../../../components/ShowsCollection';
 
 const SimilarShowsSection = () => {
   const { shows, loading } = useContext(SimilarShowsContext);
 
   return (
     <Section title="Similar TV Shows">
-      <SimilarShowsCollection shows={shows} loading={loading} />
+      <ShowsCollection shows={shows} loading={loading} scroll />
     </Section>
   );
 };
